Rename: comment 컴포넌트 폴더 board로 이동 (#63)
